feat(swiftui): wrap side-by-side diff with shared wrap_sbs_rows
