+++ shell/mac/Packages/JayJayDiffUI/Sources/JayJayDiffUI/SideBySideDiffView.swift
@@ -40,8 +40,9 @@ public struct SideBySideRepresentable: NSViewRepresentable {
 
         let left = makeContainer()
         let right = makeContainer()
-        // Per-side wrapping desyncs the two panes' visual rows from each other and
-        // from the gutters. Disable until SBS gains "wrap to tallest side" alignment.
+        // We pre-wrap rows in Rust via `wrapSbsRows` so each visual row is shorter
+        // than the pane's column count. The NSTextContainer therefore should NOT
+        // wrap on its own — that would re-wrap the pre-wrapped row and desync the panes.
         left.wrapsText = false
         right.wrapsText = false
         split.addSubview(left)
@@ -55,102 +56,11 @@ public struct SideBySideRepresentable: NSViewRepresentable {
     }
 
     public func updateNSView(_ split: NSSplitView, context: Context) {
-        guard let leftContainer = context.coordinator.leftContainer,
-              let rightContainer = context.coordinator.rightContainer,
-              let leftTV = leftContainer.textView as NSTextView?,
-              let rightTV = rightContainer.textView as NSTextView?,
-              let leftGutterTV = leftContainer.gutterTextView as DiffGutterTextView?,
-              let rightGutterTV = rightContainer.gutterTextView as DiffGutterTextView?,
-              let leftLayout = leftTV.layoutManager as? DiffLayoutManager,
-              let rightLayout = rightTV.layoutManager as? DiffLayoutManager,
-              let leftGutterLayout = leftGutterTV.layoutManager as? DiffLayoutManager,
-              let rightGutterLayout = rightGutterTV.layoutManager as? DiffLayoutManager
-        else { return }
-
         let font = NSFont(name: fontFamily, size: fontSize) ?? .monospacedSystemFont(ofSize: fontSize, weight: .regular)
         let theme = DiffColors(isDark: colorScheme == .dark)
-        let rows = buildSideBySideRows(lines: diff.lines)
-
-        let leftText = NSMutableAttributedString()
-        let rightText = NSMutableAttributedString()
-        let leftGutter = NSMutableAttributedString()
-        let rightGutter = NSMutableAttributedString()
-        var leftEntries: [DiffGutterTextView.Entry] = []
-        var rightEntries: [DiffGutterTextView.Entry] = []
-        var leftWidth: CGFloat = 0
-        var rightWidth: CGFloat = 0
-        var leftColors: [NSColor] = []
-        var rightColors: [NSColor] = []
-
-        let gutterParagraphStyle = NSMutableParagraphStyle()
-        gutterParagraphStyle.alignment = .right
-        let gutterAttrs: [NSAttributedString.Key: Any] = [
-            .font: font,
-            .foregroundColor: theme.gutterText,
-            .paragraphStyle: gutterParagraphStyle
-        ]
-        let trailingPadding: CGFloat = 10
-
-        for row in rows {
-            appendTextLine(
-                to: leftText,
-                spans: row.oldSpans,
-                style: row.oldStyle,
-                font: font,
-                theme: theme,
-                bgColors: &leftColors
-            )
-            appendTextLine(
-                to: rightText,
-                spans: row.newSpans,
-                style: row.newStyle,
-                font: font,
-                theme: theme,
-                bgColors: &rightColors
-            )
-            appendGutterLine(
-                to: leftGutter,
-                entries: &leftEntries,
-                lineNo: row.oldLineNo,
-                style: row.oldStyle,
-                attrs: gutterAttrs,
-                inset: leftGutterTV.textContainerInset.width,
-                trailingPadding: trailingPadding,
-                width: &leftWidth
-            )
-            appendGutterLine(
-                to: rightGutter,
-                entries: &rightEntries,
-                lineNo: row.newLineNo,
-                style: row.newStyle,
-                attrs: gutterAttrs,
-                inset: rightGutterTV.textContainerInset.width,
-                trailingPadding: trailingPadding,
-                width: &rightWidth
-            )
-        }
-
-        if rows.isEmpty {
-            let attrs: [NSAttributedString.Key: Any] = [
-                .font: font,
-                .foregroundColor: NSColor.secondaryLabelColor
-            ]
-            leftText.append(NSAttributedString(string: "No differences", attributes: attrs))
-            leftGutter.append(NSAttributedString(string: "\n", attributes: gutterAttrs))
-            rightGutter.append(NSAttributedString(string: "\n", attributes: gutterAttrs))
-        }
-
-        leftLayout.lineBgColors = leftColors
-        rightLayout.lineBgColors = rightColors
-        leftGutterLayout.lineBgColors = leftColors
-        rightGutterLayout.lineBgColors = rightColors
-        leftTV.textStorage?.setAttributedString(leftText)
-        rightTV.textStorage?.setAttributedString(rightText)
-        leftGutterTV.textStorage?.setAttributedString(leftGutter)
-        rightGutterTV.textStorage?.setAttributedString(rightGutter)
-        leftGutterTV.entries = leftEntries
-        rightGutterTV.entries = rightEntries
-        leftContainer.updateGutterWidth(max(52, leftWidth))
-        rightContainer.updateGutterWidth(max(52, rightWidth))
+        context.coordinator.diff = diff
+        context.coordinator.font = font
+        context.coordinator.theme = theme
+        context.coordinator.renderIfNeeded(force: true)
     }
 }
